Moved csdl loading into a utility class. Added file headers.
