+++ ODataExampleGenerator/ComponentInterfaces/IIdProvider.cs
@@ -1,4 +1,8 @@
-﻿using System;
+﻿// <copyright file="IIdProvider.cs" company="Microsoft">
+// © Microsoft. All rights reserved.
+// </copyright>
+
+using System;
 using System.Collections.Generic;
 using System.Text;
