+++ ODataExampleGenerator/GenerationParameters.cs
@@ -1,4 +1,8 @@
-﻿using System.Net.Http;
+﻿// <copyright file="GenerationParameters.cs" company="Microsoft">
+// © Microsoft. All rights reserved.
+// </copyright>
+
+using System.Net.Http;
 using System.Runtime.InteropServices.ComTypes;
 
 namespace ODataExampleGenerator
@@ -29,31 +33,7 @@ namespace ODataExampleGenerator
         public IDictionary<string, string> ChosenPrimitives { get; } =
             new Dictionary<string, string>();
 
-        public void LoadModel(string csdlFileFullPath)
-        {
-            if (!File.Exists(csdlFileFullPath))
-            {
-                throw new InvalidOperationException($"Unable to locate csdl file: {csdlFileFullPath}");
-            }
-
-            var reader = XmlReader.Create(new StringReader(File.ReadAllText(csdlFileFullPath)));
-
-            if (CsdlReader.TryParse(reader, false, out IEdmModel model, out IEnumerable<EdmError> errors))
-            {
-                this.Model = model;
-            }
-            else
-            {
-                var errorMessages = new StringBuilder();
-                foreach (var error in errors)
-                {
-                    errorMessages.AppendLine(error.ErrorMessage);
-                }
-
-                throw new InvalidOperationException(
-                    $@"Failed to read model {csdlFileFullPath}.\r\nErrors:\r\n{errorMessages}");
-            }
-        }
+        
     }
 
     public enum GenerationStyle
